测试套件重构、性能文档与README更新\n\n- 测试重构：按模块拆分 tests/gint_test.cpp，合并 coverage_gaps/int256/int512 到对应套件；新增 exceptions_overflow 目标并启用除零检查；clang-format 格式化测试代码。\n- 构建与覆盖率：CMake 注册新测试目标；Makefile 将 coverage 默认改为 gcovr（输出 HTML/XML/JSON），保留 lcov 备选。\n- 文档：新增 docs/TECH_SPEC.md；将 bench/RESULTS.md 迁移至 docs/BENCHMARKS.md，结果改表格、标注 Apple Silicon 平台并预留 x86 章节；README 更新性能摘要与文档链接（Spec/Bench）。\n- 验证：全部测试通过；bench 复测数据已同步到文档。
