+++ tests/stream_test.cpp
@@ -1,0 +1,27 @@
+#include <sstream>
+#include <gint/gint.h>
+#include <gtest/gtest.h>
+
+TEST(WideIntegerStream, Output)
+{
+    gint::integer<128, unsigned> v = 42;
+    std::ostringstream oss;
+    oss << v;
+    EXPECT_EQ(oss.str(), "42");
+}
+
+TEST(WideIntegerStream, OutputNegative)
+{
+    gint::integer<128, signed> v = -42;
+    std::ostringstream oss;
+    oss << v;
+    EXPECT_EQ(oss.str(), "-42");
+}
+
+TEST(WideIntegerStream, OutputZero)
+{
+    gint::integer<128, unsigned> v = 0;
+    std::ostringstream oss;
+    oss << v;
+    EXPECT_EQ(oss.str(), "0");
+}
